Made new layout for Print with all print listners, opening lyric print in new window
